// athome.financialVet/src/components/MeusDados.d.ts
export interface DadosProfissional {
    nome: string;
    crmv: string;
    telefone: string;
    email: string;
    pix: string;
    banco: string;
    agencia: string;
    conta: string;
    observacoesPagamento: string;
    logo?: string;
}
export declare function MeusDados(): import("react/jsx-runtime").JSX.Element;
//# sourceMappingURL=MeusDados.d.ts.map